@@ -1,9 +1,10 @@
 name: faker
-version: 1.3.0
+version: 2.0.0
 description: A library for generating fake data. faker is heavily inspired by the Python package faker and, the Ruby package ffaker.
 homepage: https://github.com/drager/faker
-author: Jesper Håkansson <jesper@jesperh.se>
 environment:
-  sdk: '>=2.0.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
+dependencies:
+  crypto: ^3.0.0
 dev_dependencies:
-  test: ^1.2.0
+  test: ^1.16.0-nullsafety.17
