@@ -3,6 +3,9 @@ library faker.example;
 import 'package:faker/faker.dart';
 
 void main() {
+  // final random = RandomGenerator(seed: 63833423);
+  // final faker = Faker.withGenerator(random);
+
   // Generate some different types of data.
   print(faker.address.country());
   print(faker.company.name());
@@ -27,10 +30,8 @@ void main() {
   // Generate random User Agent, with osName iOS
   print(faker.internet.userAgent(osName: 'iOS'));
 
-  // Generate localized Farsi text
-  var fakerFa = Faker(FakerDataProviderFa());
-
   // Generate localized Farsi lorem text
+  final fakerFa = Faker(provider: FakerDataProviderFa());
   print(fakerFa.lorem.word());
   print(fakerFa.lorem.sentence());
   print(fakerFa.lorem.words(10).join(' '));
@@ -42,6 +43,9 @@ void main() {
   print(fakerFa.person.name());
   print(fakerFa.person.fullName());
 
-
-
+  // Generate JWT(json web token)
+  print(faker.jwt.valid());
+  print(faker.jwt.valid(expiresIn: DateTime.now().add(Duration(hours: 2))));
+  print(faker.jwt.expired());
+  print(faker.jwt.custom(expiresIn: DateTime.now(), payload: {'key': 'value'}));
 }
