@@ -3,15 +3,20 @@ import 'data/address/continents.dart';
 import 'data/address/countries.dart';
 import 'data/address/country_codes.dart';
 import 'data/address/neighborhoods.dart';
+import 'data/address/states.dart';
 import 'data/address/street_suffixes.dart';
-import 'faker.dart';
+import 'person.dart';
 import 'random_generator.dart';
 
 class Address {
   static const compassDirections = ['North', 'East', 'West', 'South'];
   static const cityPrefixes = ['New', 'Lake', 'Port'];
 
-  const Address();
+  const Address(this.person);
+
+  final Person person;
+
+  RandomGenerator get random => person.random;
 
   /// Generates a zip code.
   ///
@@ -30,15 +35,15 @@ class Address {
   String city() {
     switch (random.integer(4)) {
       case 0:
-        return '${cityPrefix()} ${faker.person.firstName()}${citySuffix()}';
+        return '${cityPrefix()} ${person.firstName()}${citySuffix()}';
       case 1:
-        return '${cityPrefix()} ${faker.person.firstName()}';
+        return '${cityPrefix()} ${person.firstName()}';
       case 2:
-        return '${faker.person.firstName()}${citySuffix()}';
+        return '${person.firstName()}${citySuffix()}';
       case 3:
-        return '${faker.person.lastName()}${citySuffix()}';
+        return '${person.lastName()}${citySuffix()}';
       default:
-        return '${faker.person.lastName()}${citySuffix()}';
+        return '${person.lastName()}${citySuffix()}';
     }
   }
 
@@ -68,8 +73,8 @@ class Address {
   ///   faker.address.streetName();
   /// ```
   String streetName() => random.integer(2) == 0
-      ? '${Faker().person.lastName()} ${streetSuffix()}'
-      : '${Faker().person.firstName()} ${streetSuffix()}';
+      ? '${person.lastName()} ${streetSuffix()}'
+      : '${person.firstName()} ${streetSuffix()}';
 
   /// Generates a street address.
   ///
@@ -102,6 +107,34 @@ class Address {
   ///   faker.address.neighborhood();
   /// ```
   String neighborhood() => random.element(neighborhoods);
+
+  /// Generates a US state.
+  ///
+  /// Example:
+  /// ```dart
+  ///   faker.address.state();
+  /// ```
+  String state() => random.mapElementValue(states);
+
+  /// Generates a US state abbreviation.
+  ///
+  /// Example:
+  /// ```dart
+  ///   faker.address.stateAbbreviation();
+  /// ```
+  String stateAbbreviation() => random.mapElementKey(states);
+
+  /// Generates a Map containing a US state and it's abbreviation.
+  ///
+  /// Example:
+  /// ```dart
+  ///   faker.address.stateAsMap();
+  /// ```
+  Map<String, String> stateAsMap() {
+    String key = random.mapElementKey(states);
+    String value = states[key]!;
+    return {'state': key, 'abbreviation': value};
+  }
 
   /// Generates a country.
   ///
