@@ -2,7 +2,9 @@ import 'data/conference/conference_names.dart';
 import 'random_generator.dart';
 
 class Conference {
-  const Conference();
+  const Conference(this.random);
+
+  final RandomGenerator random;
 
   /// Generates a conference name.
   ///
