@@ -5,16 +5,15 @@ import '../data/lorem/defaults/words.dart' as word_list;
 import '../data/person/defaults/firstnames.dart' as first_names_list;
 import '../data/person/defaults/lastnames.dart' as last_names_list;
 
-class DefaultLoremDataProvider extends LoremDataProvider{
+class DefaultLoremDataProvider extends LoremDataProvider {
   @override
   List<String> sentencesList() => sentence_list.sentences;
 
   @override
   List<String> wordsList() => word_list.words;
-
 }
 
-class DefaultPersonDataProvider extends PersonDataProvider{
+class DefaultPersonDataProvider extends PersonDataProvider {
   @override
   List<String> firstNames() => first_names_list.firstnames;
 
@@ -25,6 +24,6 @@ class DefaultPersonDataProvider extends PersonDataProvider{
   List<String> prefixes() => ['Mr.', 'Mrs.', 'Ms.', 'Miss', 'Dr'];
 
   @override
-  List<String> suffixes() =>  ['Jr.', 'Sr.', 'I', 'II', 'III', 'IV', 'V', 'MD', 'DDS', 'PhD', 'DVM'];
-
+  List<String> suffixes() =>
+      ['Jr.', 'Sr.', 'I', 'II', 'III', 'IV', 'V', 'MD', 'DDS', 'PhD', 'DVM'];
 }
