@@ -1,15 +1,25 @@
 import 'dart:math';
 
-var _rng = Random();
-
-const random = RandomGenerator();
+final random = RandomGenerator();
 
 class RandomGenerator {
-  const RandomGenerator();
+  RandomGenerator({int? seed}) : _random = Random(seed);
+
+  final Random _random;
 
   /// Plucks a random element from the given [list].
   T element<T>(List<T> list) {
-    return list[_rng.nextInt(list.length)];
+    return list[_random.nextInt(list.length)];
+  }
+
+  /// Plucks a random key from the given [map].
+  T mapElementKey<T>(Map<dynamic, dynamic> map) {
+    return map.keys.elementAt(_random.nextInt(map.length));
+  }
+
+  /// Plucks a random value from the given [map].
+  T mapElementValue<T>(Map<dynamic, dynamic> map) {
+    return map.values.elementAt(_random.nextInt(map.length));
   }
 
   /// Generates a random number from the [max] value
@@ -20,13 +30,13 @@ class RandomGenerator {
     var number = <int>[];
 
     for (var i = 0; i < times; i++) {
-      number.add(_rng.nextInt(max));
+      number.add(_random.nextInt(max as int));
     }
     return number;
   }
 
-  /// Generates a random integer with the given [max] value
-  /// and to the lowest [min] value if provided.
+  /// Generates a random integer with the given [max] (exclusive) value
+  /// and to the lowest [min] (inclusive) value if provided.
   ///
   /// Example:
   ///   ```dart
@@ -35,15 +45,23 @@ class RandomGenerator {
   ///   random.integer(10, min: 5);
   ///   ```
   int integer(int max, {int min = 0}) =>
-      max == min ? max : _rng.nextInt(max - min) + min;
+      max == min ? max : _random.nextInt(max - min) + min;
+
+  /// Generates a random string of numbers with a number of characters equal to [length]
+  ///
+  /// Example
+  /// ```dart
+  /// random.numberOfLength(10);
+  /// ```
+  String numberOfLength(int length) => numbers(10, length).join();
 
   /// Generates a random boolean.
-  bool boolean() => _rng.nextBool();
+  bool boolean() => _random.nextBool();
 
   /// Generates a random decimal.
   /// Accepts a [scale] and a [min] value.
   double decimal({num scale = 1, num min = 0}) =>
-      _rng.nextDouble() * scale + min;
+      _random.nextDouble() * scale + min;
 
   /// Generates a random string with the given [max] value
   /// and to the lowest [min] value if provided.
@@ -65,9 +83,9 @@ class RandomGenerator {
   ///
   ///     random.amount((_) => random.string(15), 10, min: 5);
   ///   ```
-  List amount(fn(int i), int max, {int min = 1}) {
+  List<T> amount<T>(T Function(int i) generator, int max, {int min = 1}) {
     var length = integer(max, min: min);
-    return List.generate(length, fn);
+    return List.generate(length, generator);
   }
 
   /// Generates a random set of numbers from the given [pattern].
@@ -91,4 +109,15 @@ class RandomGenerator {
   String fromPatternToHex(List pattern) => element(pattern).splitMapJoin('#',
       onMatch: (_) =>
           numbers(16, 1).map((number) => number.toRadixString(16)).join());
+
+  /// Generated a random string from the given string of chracters.
+  ///
+  /// Example:
+  ///
+  /// ``` dart
+  ///   fromCharSet('ABCDEFGHIJKLMONPQESTUVWXYZ!?@#$$%');
+  /// ```
+  String fromCharSet(String charSet, int length) =>
+      String.fromCharCodes(Iterable.generate(
+          length, (_) => charSet.codeUnitAt(_random.nextInt(charSet.length))));
 }
