@@ -2,7 +2,9 @@ import 'data/sport/sport_names.dart';
 import 'random_generator.dart';
 
 class Sport {
-  const Sport();
+  const Sport(this.random);
+
+  final RandomGenerator random;
 
   /// Generates a sport name.
   ///
