@@ -1,17 +1,21 @@
 import 'specs/address.dart' as address;
-import 'specs/conference.dart' as conference;
 import 'specs/company.dart' as company;
+import 'specs/conference.dart' as conference;
 import 'specs/currency.dart' as currency;
+import 'specs/date.dart' as date;
 import 'specs/food.dart' as food;
 import 'specs/guid.dart' as guid;
 import 'specs/image.dart' as image;
 import 'specs/internet.dart' as internet;
 import 'specs/job.dart' as job;
+import 'specs/jwt.dart' as jwt;
 import 'specs/lorem.dart' as lorem;
 import 'specs/person.dart' as person;
 import 'specs/sport.dart' as sport;
-import 'specs/date.dart' as date;
+import 'specs/vehicle.dart' as vehicle;
+import 'specs/phone_number.dart' as phoneNumber;
 import 'specs/random_generator.dart' as random;
+import 'specs/seed.dart' as seed;
 
 void main() {
   address.main();
@@ -23,9 +27,13 @@ void main() {
   image.main();
   internet.main();
   job.main();
+  jwt.main();
   lorem.main();
   person.main();
+  phoneNumber.main();
   sport.main();
+  vehicle.main();
   date.main();
   random.main();
+  seed.main();
 }
