@@ -5,8 +5,10 @@ void main() {
   group('company', () {
     test('should be able to generate name', () {
       for (var i = 0; i < 10; i++) {
-        expect(faker.company.name(), anyOf(matches(r'^[^ ]+ [^ ]$'),
-            matches(r"^[\w-']+$"), matches(r"^[\w,' ]+[\w ]$")));
+        expect(
+            faker.company.name(),
+            anyOf(matches(r'^[^ ]+ [^ ]$'), matches(r"^[\w-']+$"),
+                matches(r"^[\w,' ]+[\w ]$")));
       }
     });
 
